FVM: Move ChainID to shared
