+++ shared/src/lib.rs
@@ -10,6 +10,7 @@ use clock::ChainEpoch;
 
 pub mod address;
 pub mod bigint;
+pub mod chainid;
 pub mod clock;
 pub mod commcid;
 pub mod consensus;
